+++ src/main/java/blackjack/card/Card.java
@@ -1,4 +1,4 @@
-package blackjack.model.card;
+package blackjack.card;
 
 public class Card {
