+++ src/main/java/blackjack/gamer/Dealer.java
@@ -1,10 +1,10 @@
-package blackjack.model.gamer;
+package blackjack.gamer;
 
-import blackjack.model.card.Card;
-import blackjack.model.card.Denomination;
-import blackjack.model.card.Hand;
-import blackjack.model.card.Shape;
-import blackjack.model.cardMachine.CardMachine;
+import blackjack.card.Card;
+import blackjack.card.Denomination;
+import blackjack.card.Hand;
+import blackjack.card.Shape;
+import blackjack.cardMachine.CardMachine;
 import java.util.Arrays;
 import java.util.List;
